Move pinned message hooks to a dedicated file
